+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to the versioning scheme outlined in the [README.md](RE
 
 - Implement faster cost tracker for default cost functions in Clarity
 - Miner will stop waiting for signatures on a block if the Stacks tip advances (causing the block it had proposed to be invalid).
+- By default, miners will wait for a new tenure to start for a configurable amount of time after receiving a burn block before 
+  submitting a block commit. This will reduce the amount of RBF transactions miners are expected to need.
 - Logging improvements:
   - P2P logs now includes a reason for dropping a peer or neighbor
   - Improvements to how a PeerAddress is logged (human readable format vs hex)